requirements.txt
matplotlib==3.8.2
numpy==2.2.5
opencv_python==4.10.0.84
Pillow==10.0.1
Pillow==11.2.1
python_Levenshtein==0.27.1
scikit_learn==1.4.0
scipy==1.15.2
torch==2.1.2
torchsummary==1.5.1
torchvision==0.16.2
tqdm==4.67.1
